requirements.txt
aiogram==3.2.0
PyYAML==6.0.1
pydantic==2.5.2
SQLAlchemy==2.0.23